requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.8.3
requests==2.31.0
pydantic==2.7.4
pydantic-settings==2.2.1
firebase-admin==6.4.0 